requirements.txt
pandas
numpy
pytorch
torchvision
torchmetrics
plotly